Move score submission logic in general out to its own Player type
