+++ osu.Game/Screens/OnlinePlay/Playlists/PlaylistsPlayer.cs
@@ -4,11 +4,8 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
-using osu.Framework.Logging;
 using osu.Framework.Screens;
 using osu.Game.Online.API;
 using osu.Game.Online.Rooms;
@@ -19,22 +16,11 @@ using osu.Game.Screens.Ranking;
 
 namespace osu.Game.Screens.OnlinePlay.Playlists
 {
-    public class PlaylistsPlayer : Player
+    public class PlaylistsPlayer : RoomSubmittingPlayer
     {
         public Action Exited;
 
-        [Resolved(typeof(Room), nameof(Room.RoomID))]
-        protected Bindable<long?> RoomId { get; private set; }
-
         protected readonly PlaylistItem PlaylistItem;
-
-        protected long? Token { get; private set; }
-
-        [Resolved]
-        private IAPIProvider api { get; set; }
-
-        [Resolved]
-        private IBindable<RulesetInfo> ruleset { get; set; }
 
         public PlaylistsPlayer(PlaylistItem playlistItem, PlayerConfiguration configuration = null)
             : base(configuration)
@@ -43,12 +29,8 @@ namespace osu.Game.Screens.OnlinePlay.Playlists
         }
 
         [BackgroundDependencyLoader]
-        private void load()
+        private void load(IBindable<RulesetInfo> ruleset)
         {
-            Token = null;
-
-            bool failed = false;
-
             // Sanity checks to ensure that PlaylistsPlayer matches the settings for the current PlaylistItem
             if (Beatmap.Value.BeatmapInfo.OnlineBeatmapID != PlaylistItem.Beatmap.Value.OnlineBeatmapID)
                 throw new InvalidOperationException("Current Beatmap does not match PlaylistItem's Beatmap");
@@ -58,30 +40,11 @@ namespace osu.Game.Screens.OnlinePlay.Playlists
 
             if (!PlaylistItem.RequiredMods.All(m => Mods.Value.Any(m.Equals)))
                 throw new InvalidOperationException("Current Mods do not match PlaylistItem's RequiredMods");
-
-            var req = new CreateRoomScoreRequest(RoomId.Value ?? 0, PlaylistItem.ID, Game.VersionHash);
-            req.Success += r => Token = r.ID;
-            req.Failure += e =>
-            {
-                failed = true;
-
-                if (string.IsNullOrEmpty(e.Message))
-                    Logger.Error(e, "Failed to retrieve a score submission token.");
-                else
-                    Logger.Log($"You are not able to submit a score: {e.Message}", level: LogLevel.Important);
-
-                Schedule(() =>
-                {
-                    ValidForResume = false;
-                    this.Exit();
-                });
-            };
-
-            api.Queue(req);
-
-            while (!failed && !Token.HasValue)
-                Thread.Sleep(1000);
         }
+
+        protected override APIRequest<APIScoreToken> CreateTokenRequestRequest() => new CreateRoomScoreRequest(RoomId.Value ?? 0, PlaylistItem.ID, Game.VersionHash);
+
+        public override APIRequest<MultiplayerScore> CreateSubmissionRequest(Score score, int token) => new SubmitRoomScoreRequest(token, RoomId.Value ?? 0, PlaylistItem.ID, score.ScoreInfo);
 
         public override bool OnExiting(IScreen next)
         {
@@ -104,31 +67,6 @@ namespace osu.Game.Screens.OnlinePlay.Playlists
             var score = base.CreateScore();
             score.ScoreInfo.TotalScore = (int)Math.Round(ScoreProcessor.GetStandardisedScore());
             return score;
-        }
-
-        protected override async Task SubmitScore(Score score)
-        {
-            await base.SubmitScore(score).ConfigureAwait(false);
-
-            Debug.Assert(Token != null);
-
-            var tcs = new TaskCompletionSource<bool>();
-            var request = new SubmitRoomScoreRequest(Token.Value, RoomId.Value ?? 0, PlaylistItem.ID, score.ScoreInfo);
-
-            request.Success += s =>
-            {
-                score.ScoreInfo.OnlineScoreID = s.ID;
-                tcs.SetResult(true);
-            };
-
-            request.Failure += e =>
-            {
-                Logger.Error(e, "Failed to submit score");
-                tcs.SetResult(false);
-            };
-
-            api.Queue(request);
-            await tcs.Task.ConfigureAwait(false);
         }
 
         protected override void Dispose(bool isDisposing)
